+++ demo/src/main/java/com/example/demo/repository/PostRepository.java
@@ -3,8 +3,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.model.Post;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
-    Optional<Post> findByPostId(Integer postId); 
+public interface PostRepository extends JpaRepository<Post, Long> {
+    Optional<Post> findByPostId(Long postId); 
     
 }
